requirements.txt
django==3.2.6
phonenumbers>=7.0.2
django-phonenumber-field==5.2.0
django-bootstrap5==2.1.1
Babel==2.9.1